定数の定義をconfig.pyに移動
