+++ app/modules/core/__init__.py
@@ -1,0 +1,4 @@
+"""
+コアモジュール
+アプリケーションの基本的な機能を提供
+"""
