+++ app/modules/optimization/__init__.py
@@ -1,0 +1,4 @@
+"""
+最適化モジュール
+アプリケーションのパフォーマンス最適化機能を提供
+"""
